backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
google-genai>=0.6.0
gunicorn==21.2.0
python-dotenv==0.21.0
pytest==8.2.2
reportlab==4.4.4
sqlalchemy==2.0.44
uvicorn==0.27.1
werkzeug==3.1.4
